pyproject.toml
[project]
name = "yahoo-symbols"
version = "0.1.8"
description = "Balzing fast yahoo symbol downloader."
authors = [
    {name = "Volker Lorrmann", email = "volker.lorrmann@gmail.com"},
]
dependencies = [
    "pydala2 @ git+https://github.com/legout/pydala2",
    "msgspec>=0.18.4",
    "typer>=0.9.0",
    "rich>=13.6.0",
    #"rocketry>=2.5.1",
    "polars[sqlalchemy]>=0.19.12",
    "yfin @ git+https://github.com/legout/yfin",
    #"pydantic==1.10.10",
    "parallel-requests @ git+https://github.com/legout/parallel-requests",
]
requires-python = ">=3.11,<3.13"
readme = "README.md"
license = {text = "MIT"}

[project.optional-dependencies]
dev = [
    "jupyterlab>=4.0.7",
    "black>=23.9.1",
    "isort>=5.12.0",
    "ipython>=8.16.1",
    "ruff>=0.1.3",
    "tabulate>=0.9.0",
    "flask>=3.0.0",
    "flask-apscheduler>=1.13.1",
]

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm.build]
includes = [
    "yahoo_symbols",
]

[tool.pdm.scripts]
#scheduler = "python -m yahoo_symbols.schedule"
downloader = "python -m yahoo_symbols.run run"
